+++ app/src/main/res/navigation/navigation_graph.xml
@@ -1,3 +1,4 @@
+<!-- 遷移グラフを定義する -->
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
